# app/swagger.yaml
openapi: 3.0.3
info:
  title: LorrySwap API - OpenAPI 3.0
  termsOfService: http://swagger.io/terms/
  contact:
    email: fernandowkd.21@itfac.mrt.ac.lk
  version: 1.0.11
servers:
  - url: https://lorryswapbackendtest2.azurewebsites.net  
...
paths:
...
  /owner/assign_vehicle_to_driver/{userId}:
    put:
      tags:
        - Owner
      summary: "Assign vehicle to driver"
      description: "This endpoint allows the owner to assign a vehicle to an existing driver, identified by the driver's user ID."
      operationId: owner.assign_vehicle_to_driver
      parameters:
        - name: userId
          in: path
          description: "The user ID of the driver to be assigned a vehicle."
          required: true
          schema:
            type: integer
      requestBody:
        description: "Data required to assign the vehicle"
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                vehicle_no:
                  type: string
                  example: "ABC1234"
      responses:
        "200":
          description: "Vehicle assigned to driver."
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Vehicle assigned to driver."
        "400":
          description: "Bad request if the vehicle number is not provided or is invalid."
        "401":
          description: "Unauthorized if the user is not authenticated."
        "403":
          description: "Forbidden if the authenticated user does not have permission to assign vehicles."
        "404":
          description: "Not found if the driver's userId does not exist in the system."
  /owner/add_vehicle/{organizationId}:
    post:
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
      tags:
        - Owner
      summary: "Add a new vehicle to the organization"
      description: "Allows the owner to add a new vehicle to their organization."
      operationId: owner.add_vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Vehicle"
      responses:
        "201":
          description: "Vehicle successfully added."
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Vehicle added successfully."
        "400":
          description: "Invalid input or missing required fields."
        "401":
          description: "Unauthorized if the user is not authenticated."
        "500":
          description: "Internal Server Error."
...
  /owner/edit_vehicle/{userId}:
    put:
      tags:
        - Owner
      summary: "Edit an existing vehicle"
      description: "Allows the owner to update details of an existing vehicle."
      operationId: owner.edit_vehicle
      parameters:
        - name: userId
          in: path
          description: "The ID of the vehicle to be edited."
          required: true
          schema:
            type: integer
      requestBody:
        description: "Data required to edit the vehicle"
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                vehicle_no:
                  type: string
                  example: "ABC1234"
  
      responses:
        "200":
          description: "Vehicle updated successfully."
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Vehicle updated successfully."
        "400":
          description: "Invalid input or missing required fields."
        "404":
          description: "Vehicle not found."
        "401":
          description: "Unauthorized if the user is not authenticated."
        "500":
          description: "Internal Server Error."
  /owner/get_organization_details/{organizationId}:
    get:
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
      tags:
        - Owner
      summary: "Get details of the organization"
      description: "Fetches details of the organization including drivers and vehicles associated with the owner."
      operationId: owner.get_organization_details
      responses:
        "200":
          description: "Organization details retrieved successfully."
          content:
            application/json:
              schema:
                type: object
                properties:
                  drivers:
                    type: array
                    items:
                      $ref: "#/components/schemas/driver"
                  vehicles:
                    type: array
                    items:
                      $ref: "#/components/schemas/Vehicle"
        "401":
          description: "Unauthorized if the user is not authenticated."
        "404":
          description: "Organization details not found."
        "500":
          description: "Internal Server Error."
  /owner/get_drivers_by_organization/{organizationId}:
    get:
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
      tags:
        - Owner
      summary: "Get drivers by organization id"
      description: "Fetches drivers of the organization associated with the owner."
      operationId: owner.get_drivers_by_organization
      responses:
        "200":
          description: "Driver details by organizationId retrieved successfully."
          content:
            application/json:
              schema:
                type: object
                properties:
                  drivers:
                    type: array
                    items:
                      $ref: "#/components/schemas/driver"
        "401":
          description: "Unauthorized if the user is not authenticated."
        "404":
          description: "Organization details not found."
        "500":
          description: "Internal Server Error."
  /owner/get_vehicles_by_organization/{organizationId}:
    get:
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: integer
      tags:
        - Owner
      summary: "Get vehicles by organization id"
      description: "Fetches vehicle detais of the organization associated with the owner."
      operationId: owner.get_vehicles_by_organization
      responses:
        "200":
          description: "Driver details by organizationId retrieved successfully."
          content:
            application/json:
              schema:
                type: object
                properties:
                  drivers:
                    type: array
                    items:
                      $ref: "#/components/schemas/Vehicledetail"
        "401":
          description: "Unauthorized if the user is not authenticated."
        "404":
          description: "Organization details not found."
        "500":
          description: "Internal Server Error."
  /owner/delete_driver/{userId}:
    put:
      tags:
        - Owner
      summary: "Delete driver"
      description: "delete driver from owner's organization"
      operationId: owner.delete_driver
      parameters:
        - name: userId
          in: path
          description: "The ID of the vehicle to be edited."
          required: true
          schema:
            type: integer
      requestBody:
        description: "Data required to edit the vehicle"
        required: true
        content:
          application/json:
            schema:
                type: array
                items:
                  type: string
  
      responses:
        "200":
          description: "Vehicle updated successfully."
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Vehicle updated successfully."
        "400":
          description: "Invalid input or missing required fields."
        "404":
          description: "Vehicle not found."
        "401":
          description: "Unauthorized if the user is not authenticated."
        "500":
          description: "Internal Server Error."
  /owner/delete_vehicle/{Vehicleno}:
    delete:
      tags:
        - Owner
      summary: Delete a vehicle
      description: Deletes a vehicle based on the vehicle no.
      operationId: owner.delete_vehicle
      parameters:
        - in: path
          name: Vehicleno
          required: true
          description: The ID of the vehicle will be deleted.
          schema:
            type: string
      responses:
        '200':
          description: Vehicle deleted successfully
        '404':
          description: vehicle not found
  /owner/get_driver_by_userId/{userId}:
    get:
      tags:
        - Owner
      summary: Get driver details by userId
      description: Get driver details by userId
      operationId: owner.get_driver_by_userId
      parameters:
        - in: path
          name: userId
          required: true
          description: The ID of the user
          schema:
            type: integer
      responses:
        '200':
          description: Driver details userId retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 123
                  fname:
                    type: string
                    example: kamal
                  lname:
                    type: string
                    example: senarath
                  phone:
                    type: string
                    example: 0719563248
                  nic:
                    type: string
                    example: 200135694586
                  age:
                    type: string
                    example: 25
                  gender:
                    type: string
                    example: Male
                  email:
                    type: string
                    example: 123abc@gmail.com
                  residence:
                    type: string
                    example: Ratnapura
                  Vehicleno:
                    type: string
                    example: "ABC123"

        "401":
          description: "Unauthorized if the user is not authenticated."
        "404":
          description: "driver details not found."
        "500":
          description: "Internal Server Error."
...
  /driver_details:
    get:
      tags:
        - Admin Dashboard
      summary: Get Driver Details
      operationId: AdminDashBoard.driver_details
      description: Get all details for admin web app
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Get_driver_details'
        '400':
          description: Invalid data supplied
...
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: auth.bearer_info_func
  schemas:
...
    Vehicle:
      type: object
      properties:
        Vehicleno:
          type: string
          example: BDP4515
        brand:
          type: string
          example: TOYOTA
        model:
          type: string
          format: Axios
        type:
          type: string
          example: Coverd
        length:
          type: number
          example: 10
          description: "in meters"
        width:
          type: number
          example: 3
          description: "in meters"
        height:
          type: number
          example: 4
          description: "in meters"
        tonnage:
          type: number
          example: 7 
          description: "in metricton"
      required:
        - Vehicleno
        - brand
        - model
        - type
        - length
        - width
        - height
        - tonnage
...
    driver:
      type: object
      properties:
        id:
          type: integer
        fname:
          type: string
        lname:
          type: string
        phone:
          type: string
      required:
        - fname
        - lname
        - phone
...